Refactor data types from long long to int for improved performance and memory usage in game logic and segment tree implementation. Update related functions and variables accordingly to ensure consistency and correctness.

+++ 线段树AI代码（助理解）（争取2026.7.18之前速通）.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 const int N = 1e5 + 10;
 
-long long tr[4 * N];   // 线段树存区间和
-long long lazy[4 * N]; // 懒标记：区间加的延迟值
-int a[N];              // 原数组
+int tr[4 * N];   // 线段树存区间和
+int lazy[4 * N]; // 懒标记：区间加的延迟值
+int a[N];        // 原数组
 
 // 上传：用左右儿子更新当前节点
 void pushup(int u) {
@@ -42,7 +42,7 @@ void build(int u, int l, int r) {
 }
 
 // 区间修改：[l,r] 加 val，当前节点u区间[l,r]
-void update(int u, int l, int r, int l, int r, long long val) {
+void update(int u, int l, int r, int l, int r, int val) {
     if (l <= l && r <= r) { // 当前区间完全被覆盖，打懒标记
         tr[u] += val * (r - l + 1);
         lazy[u] += val;
@@ -56,11 +56,11 @@ void update(int u, int l, int r, int l, int r, long long val) {
 }
 
 // 区间查询 [l,r] 和
-long long query(int u, int l, int r, int l, int r) {
+int query(int u, int l, int r, int l, int r) {
     if (l <= l && r <= r) return tr[u];
     pushdown(u, l, r);
     int mid = (l + r) / 2;
-    long long res = 0;
+    int res = 0;
     if (l <= mid) res += query(u << 1, l, mid, l, r);
     if (r > mid) res += query(u << 1 | 1, mid + 1, r, l, r);
     return res;
@@ -75,7 +75,7 @@ int main() {
     build(1, 1, n);
     while (m--) {
         int op, l, r;
-        long long x;
+        int x;
         cin >> op;
         if (op == 1) {
             // 区间加
